Make the agent reusable in multiple main()s
To simplify building of the agent and to make the root and
start cobra.Command objects reusable, apply the same treatment
to the agent that the backend got with the backend/cmd
package. The agent's can be somewhat simpler given the less
complicated initialization that has to happen.
