+++ cmd/sensu-agent/logger.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/sirupsen/logrus"
+
+var logger = logrus.WithFields(logrus.Fields{
+	"component": "agent",
+})
