+++ cmd/sensu-agent/main.go
@@ -1,18 +1,9 @@
 package main
 
-import (
-	"github.com/spf13/cobra"
-)
-
-var (
-	rootCmd = &cobra.Command{
-		Use:   "sensu-agent",
-		Short: "sensu agent",
-	}
-)
+import "github.com/sensu/sensu-go/agent/cmd"
 
 func main() {
-	if err := rootCmd.Execute(); err != nil {
+	if err := cmd.Execute(); err != nil {
 		logger.WithError(err).Fatal("error executing sensu-agent")
 	}
 }
